Create utils module and refactor analysis scripts
